# Rewrite ngram implementation with an eye for cpp11 performance